Added the current python test file
This should work right out of the gate, assuming you have the libraries installed and dataset in hand. Make sure the dataset it's looking for matches the name of the file you have. Adding graph based learning should be very simple, once you have the functions it wants created internally.