Add files via upload
The process_break_curve_data_arc.py file is a modification of the regular process_break_curve.py file that appends the endpoints to the xyz coordinate vectors. They aren't compatible, and we should look at how to best adapt these files into a single thing to reduce duplication. 

To create the arc lengths/angles, run the files in this order:
1. mesh_stats.py <- This requires all the actual meshes
2. process_break_curve_data_arc.py <- This requires the ridgelengths csv
3. order_all_breaks.py <- This requires the modifed pickle file
4. arcvars.py <- This requires the mesh_stats csv