Uploading the Moclan replication file
So, my computer keeps crashing, and while I need to get that figured out, I'm uploading the current version here so I can potentially use a different workstation. The interface isn't the best, as it relies heavily on commenting and uncommenting code to change parameters. That will be fixed in a later version. 

Results are pending me actually being able to run the code for long enough to get results.