Made train-test permute order
Because of our voting scheme, we need to permute the order of the train and test datasets. Check the function for a more detailed explanation, but basically, under certain circumstances, the order of the input data might impact the accuracy of the test.

Additionally, its just good practice (even if it's not used) to ensure that the order of the input data doesn't matter over replications.