Refactoring and adding new tests
We've changed how tests are run in this case. You now need to define an "options" dictionary that tells the dataset generating function what to bring. It's not pretty, but it is explicit. Because we're using a loop to run through the tests, I couldn't think of a good way to make it cleaner. My defaultdict idea (that would save you from having to define the things you don't want to change) went nowhere, fast. 

Added a new compilation function to ml_test.py, which does all our fancy math stuff and creates the lists for saving. The save function now also appends the current time and date to prevent file overwriting (which I've had happen to me. Lots).