Added F1 scores
Added F1 scores to the main ml_test.py file. They save under the summary files for each test, both vote and component. In the cases where those are the same, the scores are the same. 

I planned to add in a specific scalar for our categorical variables (standard scalar treats them as if they were floats), but that turned out to be a real mess to implement, and very fragile. It wouldn't have been better (it wouldn't have changed the scores), per-se, either, just more clear for us. Just standard scaling will do fine for those algorithms that like it.